refactor: move AssignmentDto in Application on level up
